+++ app/src/styles/globals.css
